requirements/optional.txt
django-compressor
django-filebrowser
django-grappelli
django-extensions
django-debug-toolbar